+++ src/engine/textrender.h
@@ -45,7 +45,7 @@ public:
 	virtual void TextColor(float r, float g, float b, float a) = 0;
 	virtual void Text(void *pFontSetV, float x, float y, float Size, const char *pText, int MaxWidth) = 0;
 	virtual float TextWidth(void *pFontSetV, float Size, const char *pText, int Length) = 0;
-	virtual float TextLineCount(void *pFontSetV, float Size, const char *pText, int LineWidth) = 0;
+	virtual int TextLineCount(void *pFontSetV, float Size, const char *pText, float LineWidth) = 0;
 };
 
 class IEngineTextRender : public ITextRender
